+++ app/controllers/api/educations_controller.rb
@@ -1,4 +1,7 @@
 class Api::EducationsController < ApplicationController
+
+  before_action :authenticate_student, only: [:create, :update, :destroy]
+
   def index
     @educations = Education.all
     render "index.json.jb"
@@ -26,22 +29,31 @@ class Api::EducationsController < ApplicationController
   end
 
   def update
-    @education = Education.find_by(id: params[:id])
-
-    @education.start_date = params[:start_date] || @education.start_date
-    @education.end_date = params[:end_date] || @education.end_date
-    @education.degree = params[:degree] || @education.degree
-    @education.university = params[:university] || @education.university
-    @education.details = params[:details] || @education.details
-   
     
-    @education.save
-    render "show.json.jb"
+    @education = Education.find_by(id: params[:id])
+    if current_student.id == @education.student_id
+      @education.start_date = params[:start_date] || @education.start_date
+      @education.end_date = params[:end_date] || @education.end_date
+      @education.degree = params[:degree] || @education.degree
+      @education.university = params[:university] || @education.university
+      @education.details = params[:details] || @education.details
+     
+      
+      @education.save
+      render "show.json.jb"
+    else
+      render json: {error: "Education can only be edited by original owners."}, status: :unauthorized
+    end  
+
   end
 
   def destroy
     education = Education.find_by(id: params[:id])
-    education.destroy
-    render json: {message: "Education successfully destroyed!"}
+    if current_student.id == education.student_id
+      education.destroy
+      render json: {message: "Education successfully destroyed!"}
+    else
+      render json: {error: "Education can only be deleted by original owners."}, status: :unauthorized
+    end  
   end
 end
